+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - BREAKING: updated minimum Ruby version required to 2.7.
 - BREAKING: removed `Hash#map_values` and `Hash#map_keys`.
 - BREAKING: removed `Proc#|`.
+- BREAKING: removed `Object#thru` and renamed `#thru_if`, `#thru_unless` to `#then_if`, `#then_unless`; renamed their module also.
 - BREAKING: `Exception#details=` and `Exception#details=` now enforce their argument to be a Hash.
 
 ## [0.2.7] - 2021-06-26